requirements.txt
chatterbox-tts==0.1.2
openai==1.109.1
telepot==12.7
torch==2.6.0
torchaudio==2.6.0